+++ Modules/CourseBatch/Http/Controllers/CourseBatchController.php
@@ -108,6 +108,10 @@ class CourseBatchController extends Controller
     {
         //
         $coursebatch = CourseBatch::find($id);
+        if($coursebatch->Registrations){
+           return redirect()->route('coursebatch_list')->With('prohibited','123'); 
+        }
+
         if($coursebatch->delete())
             return redirect()->route('coursebatch_list')->with('deleted','course Updated successfully');
         else
